# backend/api/company/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Company",
        "description": "Company management microservice documentation",
        "version": "1.0.0",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "agendai.suporte@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:3001/",
    "basePath": "/company",
    "paths": {
        "/company": {
            "post": {
                "tags": [
                    "Company"
                ],
                "description": "Create a new company",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestCreateCompany"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseMessage"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update company information",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestUpdateCompany"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company information updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseMessage"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "get": {
                "tags": [
                    "Company"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Get company information",
                "responses": {
                    "200": {
                        "description": "Company information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseCompany"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company/all-companies/{companyId}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "description": "Get all companies or a specific company by ID",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID of the company to retrieve"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Companies retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseCompanyList"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company/{companyId}": {
            "delete": {
                "tags": [
                    "Company"
                ],
                "description": "Delete a company by ID",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID of the company to be deleted"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseMessage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "RequestCreateCompany": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the company",
                        "example": "Tech Solutions"
                    },
                    "address": {
                        "type": "string",
                        "description": "Address of the company",
                        "example": "123 Tech Street"
                    },
                    "email": {
                        "type": "string",
                        "description": "Email of the company",
                        "example": "contact@techsolutions.com"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Phone number of the company",
                        "example": "123-456-7890"
                    }
                }
            },
            "RequestUpdateCompany": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the company",
                        "example": "Tech Solutions"
                    },
                    "address": {
                        "type": "string",
                        "description": "Address of the company",
                        "example": "123 Tech Street"
                    },
                    "email": {
                        "type": "string",
                        "description": "Email of the company",
                        "example": "contact@techsolutions.com"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Phone number of the company",
                        "example": "123-456-7890"
                    }
                }
            },
            "ResponseCompany": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID of the company",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the company",
                        "example": "Tech Solutions"
                    },
                    "address": {
                        "type": "string",
                        "description": "Address of the company",
                        "example": "123 Tech Street"
                    },
                    "email": {
                        "type": "string",
                        "description": "Email of the company",
                        "example": "contact@techsolutions.com"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Phone number of the company",
                        "example": "123-456-7890"
                    },
                    "createdBy": {
                        "type": "integer",
                        "description": "ID of the user who created the company",
                        "example": 1
                    }
                }
            },
            "ResponseCompanyList": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/ResponseCompany"
                }
            },
            "ResponseMessage": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Response message",
                        "example": "Operation successful"
                    }
                }
            }
        }
    }
}